Pad: Merge PadConfig/PadManager/PadMacros
Faster compile time.
