+++ pcsx2-qt/SetupWizardDialog.cpp
@@ -15,7 +15,7 @@
 
 #include "PrecompiledHeader.h"
 
-#include "pcsx2/SIO/Pad/PadConfig.h"
+#include "pcsx2/SIO/Pad/Pad.h"
 #include "QtHost.h"
 #include "QtUtils.h"
 #include "SettingWidgetBinder.h"
