--- pcsx2/SIO/Pad/PadMacros.cpp
@@ -1,90 +1,0 @@
-/*  PCSX2 - PS2 Emulator for PCs
- *  Copyright (C) 2002-2023  PCSX2 Dev Team
- *
- *  PCSX2 is free software: you can redistribute it and/or modify it under the terms
- *  of the GNU Lesser General Public License as published by the Free Software Found-
- *  ation, either version 3 of the License, or (at your option) any later version.
- *
- *  PCSX2 is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- *  without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
- *  PURPOSE.  See the GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License along with PCSX2.
- *  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#include "PrecompiledHeader.h"
-
-#include "SIO/Pad/PadMacros.h"
-
-#include "SIO/Pad/PadManager.h"
-#include "SIO/Pad/PadBase.h"
-
-namespace Pad
-{
-std::array<std::array<MacroButton, NUM_MACRO_BUTTONS_PER_CONTROLLER>, NUM_CONTROLLER_PORTS> s_macro_buttons;
-}
-
-void Pad::ClearMacros()
-{
-	s_macro_buttons = {};
-}
-
-Pad::MacroButton& Pad::GetMacroButton(u32 pad, u32 index)
-{
-	return s_macro_buttons[pad][index];
-}
-
-void Pad::SetMacroButtonState(u32 pad, u32 index, bool state)
-{
-	if (pad >= Pad::NUM_CONTROLLER_PORTS || index >= NUM_MACRO_BUTTONS_PER_CONTROLLER)
-		return;
-
-	MacroButton& mb = s_macro_buttons[pad][index];
-	if (mb.buttons.empty() || mb.trigger_state == state)
-		return;
-
-	mb.toggle_counter = mb.toggle_frequency;
-	mb.trigger_state = state;
-	if (mb.toggle_state != state)
-	{
-		mb.toggle_state = state;
-		ApplyMacroButton(pad, mb);
-	}
-}
-
-void Pad::ApplyMacroButton(u32 controller, const Pad::MacroButton& mb)
-{
-	const float value = mb.toggle_state ? mb.pressure : 0.0f;
-	PadBase* const pad = Pad::GetPad(controller);
-
-	for (const u32 btn : mb.buttons)
-		pad->Set(btn, value);
-}
-
-void Pad::UpdateMacroButtons()
-{
-	for (u32 pad = 0; pad < Pad::NUM_CONTROLLER_PORTS; pad++)
-	{
-		for (u32 index = 0; index < NUM_MACRO_BUTTONS_PER_CONTROLLER; index++)
-		{
-			Pad::MacroButton& mb = s_macro_buttons[pad][index];
-
-			if (!mb.trigger_state || mb.toggle_frequency == 0)
-			{
-				continue;
-			}
-
-			mb.toggle_counter--;
-
-			if (mb.toggle_counter > 0)
-			{
-				continue;
-			}
-
-			mb.toggle_counter = mb.toggle_frequency;
-			mb.toggle_state = !mb.toggle_state;
-			ApplyMacroButton(pad, mb);
-		}
-	}
-}
