--- pcsx2/SIO/Pad/PadMacros.h
@@ -1,44 +1,0 @@
-/*  PCSX2 - PS2 Emulator for PCs
- *  Copyright (C) 2002-2023  PCSX2 Dev Team
- *
- *  PCSX2 is free software: you can redistribute it and/or modify it under the terms
- *  of the GNU Lesser General Public License as published by the Free Software Found-
- *  ation, either version 3 of the License, or (at your option) any later version.
- *
- *  PCSX2 is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- *  without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
- *  PURPOSE.  See the GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License along with PCSX2.
- *  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#pragma once
-
-#include "SIO/Pad/PadTypes.h"
-
-#include <vector>
-#include <array>
-
-namespace Pad
-{
-	struct MacroButton
-	{
-		std::vector<u32> buttons; ///< Buttons to activate.
-		float pressure; ///< Pressure to apply when macro is active.
-		u32 toggle_frequency; ///< Interval at which the buttons will be toggled, if not 0.
-		u32 toggle_counter; ///< When this counter reaches zero, buttons will be toggled.
-		bool toggle_state; ///< Current state for turbo.
-		bool trigger_state; ///< Whether the macro button is active.
-	};
-
-	// Number of macro buttons per controller.
-	static constexpr u32 NUM_MACRO_BUTTONS_PER_CONTROLLER = 16;
-
-	// Sets the state of the specified macro button.
-	void ClearMacros();
-	MacroButton& GetMacroButton(u32 pad, u32 index);
-	void SetMacroButtonState(u32 pad, u32 index, bool state);
-	void ApplyMacroButton(u32 controller, const MacroButton& mb);
-	void UpdateMacroButtons();
-};
